added ownership check

+++ backend/controllers/postController.js
@@ -15,40 +15,51 @@ const createPost = async (req, res) => {
 };
 
 
-const getAllPosts=async(req,res) =>{
-  const posts=await Post.find().populate('author','name') 
-  res.json({posts})
+const getAllPosts = async (req, res) => {
+  const posts = await Post.find().populate('author', 'name')
+  res.json({ posts })
 }
 
-const getPostById=async(req,res) =>{
-  const post=await Post.findById(req.params.id).populate('author','name');
-  if(!post){
-    return res.status(404).json({message:'Post not found'})
+const getPostById = async (req, res) => {
+  const post = await Post.findById(req.params.id).populate('author', 'name');
+  if (!post) {
+    return res.status(404).json({ message: 'Post not found' })
   }
-  res.json({post})
+  res.json({ post })
 
-}
-
-
-const updatePost=async(req,res) =>{
-  const post=await Post.findById(req.params.id);
-  if(!post){
-    return res.status(404).json({message:'Post not found'})
-  }
-  post.title=req.body.title;
-  post.content=req.body.content;
-  const updatedPost=await post.save();
-  res.json({message:'Post updated successfully',updatedPost})
 }
 
 
-const deletePost=async(req,res) =>{
-  const post=await Post.findById(req.params.id);
-  if(!post){
-    return res.status(404).json({message:'Post not found'})
+const updatePost = async (req, res) => {
+  const post = await Post.findById(req.params.id);
+  if (!post) {
+    return res.status(404).json({ message: 'Post not found' })
   }
+
+  //ownership check
+  if (post.author.toString() !== req.user.id) {
+    return res.status(403).json({ message: 'You are not authorized to update this post' })
+  }
+
+  post.title = req.body.title;
+  post.content = req.body.content;
+  const updatedPost = await post.save();
+  res.json({ message: 'Post updated successfully', updatedPost })
+}
+
+
+const deletePost = async (req, res) => {
+  const post = await Post.findById(req.params.id);
+  if (!post) {
+    return res.status(404).json({ message: 'Post not found' })
+  }
+  //ownership check
+  if (post.author.toString() !== req.user.id) {
+    return res.status(403).json({ message: 'You are not authorized to delete this post' })
+  }
+
   await Post.findByIdAndDelete(req.params.id);
-  res.json({message:'Post deleted successfully'})
+  res.json({ message: 'Post deleted successfully' })
 }
 
-export {createPost, getAllPosts, getPostById, updatePost, deletePost}
+export { createPost, getAllPosts, getPostById, updatePost, deletePost }
